requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-debug-toolbar==4.1.0
django-allauth==0.54.0
drf-spectacular==0.26.3
django-ckeditor==6.6.1
django-filter==23.2
psycopg2==2.9.6
Pillow==10.0.0
coverage==7.2.7
celery[redis]==5.3.4
flower==2.0.0
